audio conversion struggles

+++ kurtis_mlx/utils/stt.py
@@ -12,10 +12,15 @@ def transcribe(audio_np, stt_model_name, sample_rate=16000):
     if sample_rate != 16000:
         resampling_factor = 16000 / sample_rate
         num_samples = int(len(audio_np) * resampling_factor)
-        audio_resampled = resample(audio_np, num_samples).astype(np.int16)
+
+        # Do NOT cast to int16. resample() outputs a float.
+        audio_resampled = resample(audio_np, num_samples)
     else:
         audio_resampled = audio_np
 
+    # This will now correctly normalize:
+    # 1. The new 16kHz resampled float array (from 8kHz)
+    # 2. Or the original 16kHz int16 array (in non-SIP mode)
     return mlx_whisper.transcribe(
         audio_resampled.astype(np.float32) / 32768.0,
         fp16=False,
